back/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.3
depocs==1.0.1
Django==4.2.4
django-appconf==1.0.5
django-rest-framework==0.1.0
django-sslserver==0.22
djangorestframework==3.14.0
drf-extra-fields==3.7.0
idna==3.4
mccabe==0.7.0
nulltype==2.3.1
Pillow==10.0.0
plaid-python==14.0.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.12.0
sqlparse==0.4.4
stripe==5.4.0
urllib3==2.0.3
